requirements.txt
pdfplumber>=0.11.0
python-docx>=1.2.0
pyyaml>=6.0
pytesseract>=0.3.10
pdf2image>=1.17.0
beautifulsoup4>=4.12.0
flask>=3.0
gunicorn>=22.0
